Change segmentation name and bytecode and rewrite unittest
