+++ hbctool/hbc/hbc59/test.py
@@ -1,14 +1,15 @@
-import hbc as hbcl
+from hbctool import hbc as hbcl, hasm
 from .translator import assemble, disassemble
 import unittest
 import re
 import pathlib
+import json
 
 basepath = pathlib.Path(__file__).parent.absolute()
 
-class TestHBC74(unittest.TestCase):
+class TestHBC59(unittest.TestCase):
     def __init__(self, *args, **kwargs):
-        super(TestHBC74, self).__init__(*args, **kwargs)
+        super(TestHBC59, self).__init__(*args, **kwargs)
         self.hbc = hbcl.load(open(f"{basepath}/example/index.android.bundle", "rb"))
         # self.objdump = open(f"{basepath}/example/objdump.out", "r").read()
         self.pretty = open(f"{basepath}/example/pretty.out", "r").read()
@@ -40,7 +41,7 @@ class TestHBC74(unittest.TestCase):
             # self.assertEqual(funcHeader["offset"], int(target_offset, 16))
     
     def test_get_string(self):
-        target_strings = re.findall(r"[is][0-9]+\[([UTFASCI16-]+), ([0-9]+)..([0-9-]+)\].*?:\s?(.*)", self.pretty)
+        target_strings = re.findall(r"[isp][0-9]+\[([UTFASCI16-]+), ([0-9]+)..([0-9-]+)\].*?:\s?(.*)", self.pretty)
         stringCount = self.hbc.getStringCount()
 
         self.assertEqual(stringCount, len(target_strings))
@@ -69,3 +70,29 @@ class TestHBC74(unittest.TestCase):
             _, _, _, _, bc, _ = self.hbc.getFunction(i, disasm=False)
 
             self.assertEqual(assemble(disassemble(bc)), bc)
+class TestParser59(unittest.TestCase):
+    def test_hbc(self):
+        f = open(f"{basepath}/example/index.android.bundle", "rb")
+        hbc = hbcl.load(f)
+        f.close()
+        f = open("/tmp/hbctool_test.android.bundle", "wb")
+        hbcl.dump(hbc, f)
+        f.close()
+
+        f = open("hbc/hbc59/example/index.android.bundle", "rb")
+        a = f.read()
+        f.close()
+        f = open("/tmp/hbctool_test.android.bundle", "rb")
+        b = f.read()
+        f.close()
+
+        self.assertEqual(a, b)
+
+    def test_hasm(self):
+        f = open(f"{basepath}/example/index.android.bundle", "rb")
+        a = hbcl.load(f)
+        f.close()
+        hasm.dump(a, "/tmp/hbctool_test", force=True)
+        b = hasm.load("/tmp/hbctool_test")
+
+        self.assertEqual(json.dumps(a.getObj()), json.dumps(b.getObj()))
