+++ hbctool/metadata.py
@@ -1,7 +1,7 @@
 package = 'hbctool'
 project = 'hbctool'
 project_no_spaces = project.replace(' ', '')
-version = '0.1.2'
+version = '0.1.3'
 description = ('A command-line interface for disassembling and assembling'
                'the Hermes Bytecode')
 authors = ['bongtrop']
